pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agora-telegram"
version = "2.0.0"
description = "AI圆桌会议 Telegram Bot - 支持多AI讨论、智能路由"
requires-python = ">=3.9"
dependencies = [
    "python-telegram-bot>=20.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
agora = "agora.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["agora*"]
